--- test/testhelp.js
@@ -1,5 +1,0 @@
-require('babel-register')({
-  presets: ['env'],
-});
-
-module.exports = require('./test.js');
